chore: use changeset validation in all controllers
chore: change map anotation in the pattern matching

chore: finally embedded_schemas

+++ apps/banking_api/lib/banking_api/accounts/inputs/transaction.ex
@@ -8,7 +8,7 @@ defmodule BankingApi.Accounts.Inputs.Transaction do
   embedded_schema do
     field :from, :string
     field :to, :string
-    field :value, :decimal
+    field :value, :integer
   end
 
   def changeset(model \\ %__MODULE__{}, params) do
